extract_features/requirements.txt
numpy
matplotlib
torch
torchvision 
tqdm
imutils
pillow
opencv-python
open3d